Hack to deal with relationships (Jam->Ember->Jam)
This allows us to specify the Id as text in the API payload (to or from), but treat it as a relationship field in ember. This feels a bit hacky; suggestions welcome.

The issue is that the serializer won't serialize an attribute field with the same name as an ember relationship. We use an intermediate Id field on the model (not present in the server schema) to:

a) Allow us to fetch the underlying related record ID as a string without the need for a second server call (though see https://github.com/emberjs/data/issues/2705 )

b) Allow us to send the related record ID as an attribute, as text, to the server- even though there is a relationship by the same name