Remove relationAttrs hackery
Now that we've simplified our schemas, we don't need as many manually created relationships. Remove the elements of this code that had descended into the black arts. Related records are not fetched by manually declared computed properties, and schema fields that track the ID of a related record should be renamed `<itemId>`, not `<item>`