Better experiment.thumbnail getting/setting
When an experiment record loads and it has a thumbnailId, eagerly load
the thumbnail record. Once loaded, that record is cached behind the
'thumbnail' computed.

If the thumbnailId is not set, a new record needs to be created when
trying to set the thumbnail, and the experiment record needs to be
updated with the new thumbnailId