A better way of fetching past sessions
Use a route mixin to make the exp-player route load until all of the
needed data is ready and successfully saved. This allows us to leverage
the loading substate of that route to give better feedback to the user,
and give better confidence that once the player is loaded it has
everything it needs to run.