Improve cardsort front-end validation
Currently, the "continue" button is disabled
until all cards are sorted, but it is possible to
remove the CSS class and bypass validation.
Check if all cards have been sorted in the
"nextPage" and "continue" actions so that even if
it is possible to click continue, the check still
happens.