Clear unsaved cardsort1 data
If a user sorts cards on the cardsort1 page
and then clicks the "Previous Page" button to
go to the previous page, clear the sorted cards.

Then, when they click "Continue" to go back to the
cardsort page it will display 90 cards to be sorted
at the top, and three empty buckets to sort the cards
into.