Only load data if frame changes
The didReceiveAttrs hook gets called when
a component is initially rendered, and
calls loadData to deserialize frame
data. Sometimes didReceiveAttrs gets
called again (when the component
re-renders) when the frame has not changed,
which overwrites new data with deserialized
data. Only call loadData if the frameIndex
has changed between the old and newAttrs.